Refactor extract printing results to a separate step
