+++ src/electionguard_cli/e2e_steps/__init__.py
@@ -2,10 +2,12 @@ from electionguard_cli.e2e_steps.key_ceremony_step import KeyCeremonyStep
 from electionguard_cli.e2e_steps.election_builder_step import ElectionBuilderStep
 from electionguard_cli.e2e_steps.submit_votes_step import SubmitVotesStep
 from electionguard_cli.e2e_steps.decrypt_step import DecryptStep
+from electionguard_cli.e2e_steps.print_results_step import PrintResultsStep
 
 __all__ = [
     "ElectionBuilderStep",
     "KeyCeremonyStep",
     "SubmitVotesStep",
-    "DecryptStep"
+    "DecryptStep",
+    "PrintResultsStep"
 ]
